+++ amd/src/drop-snippet-shortcode-maker.js
@@ -16,7 +16,6 @@
 /**
  * Drop snippet shortcode maker module.
  *
- * @package    block_stash
  * @copyright  2018 Frédéric Massart
  * @author     Frédéric Massart <fred@branchup.tech>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later