+++ amd/src/trade-item-dialogue.js
@@ -16,7 +16,6 @@
 /**
  * Show item detail.
  *
- * @package    block_stash
  * @copyright  2016 Adrian Greeve <adriangreeve.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -30,7 +29,9 @@ define([
     /**
      * Detail dialogue class.
      *
-     * @param {Number} itemId The item ID.
+     * @param {Number} courseid The course ID.
+     * @param {String} type The context type
+     * @param {String} title The tile of the dialogue
      */
     function Dialogue(courseid, type, title) {
         this.setTitle(title);